# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Uploy API
  version: 0.2.0
  description: Uploy deployment platform API
paths:
...
  /api/projects/{id}/services:
    get:
      operationId: listProjectServices
      summary: List services in a project
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of services for the project, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/services:
    post:
      operationId: createService
      summary: Create a new service
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Container name already in use on this server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listServices
      summary: List services in workspace
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/services/{id}:
    get:
      operationId: getService
      summary: Get service details
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateService
      summary: Update service configuration
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceRequest"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteService
      summary: Delete a service
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Service deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: >-
            The container could not be removed from the server, so the service
            was kept. Retry once the server is reachable again.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        code:
          type: string
          description: >
            Machine-readable failure stage. Present on SSH probe failures so a
            client can offer the right remedy: the three stages need entirely
            different fixes, and the human message alone cannot be branched on.
          enum:
            - unreachable
            - key_rejected
            - key_invalid
            - session_failed
            - docker_missing
...
    CreateServiceRequest:
      type: object
      required:
        - name
        - image
        - container_name
        - container_port
        - server_id
        - environment_id
      properties:
        name:
          type: string
        image:
          type: string
        container_name:
          type: string
        container_port:
          type: integer
          description: Container port the image listens on.
        host_port:
          type: integer
          description: >-
            Host port the service is published on. Omit to keep the service
            internal — reachable by other services on the uploy network and by
            nothing outside the machine, which is what a database wants. 80 and
            443 are rejected: they belong to the Uploy proxy. Ignored while the
            service has a domain, since that traffic goes through the proxy.
        server_id:
          type: string
        environment_id:
          type: string
        kind:
          type: string
          default: application
          enum:
            - application
    UpdateServiceRequest:
      type: object
      required:
        - name
        - image
        - container_name
        - container_port
        - server_id
      properties:
        name:
          type: string
        image:
          type: string
        container_name:
          type: string
        container_port:
          type: integer
          description: Container port the image listens on.
        host_port:
          type: integer
          description: >-
            Host port the service is published on. Omit to keep the service
            internal — reachable by other services on the uploy network and by
            nothing outside the machine, which is what a database wants. 80 and
            443 are rejected: they belong to the Uploy proxy. Ignored while the
            service has a domain, since that traffic goes through the proxy.
        server_id:
          type: string
    ServiceResponse:
      type: object
      required:
        - id
        - name
        - image
        - container_name
        - container_port
        - server_id
        - kind
        - project_id
        - environment_id
        - created_at
        - updated_at
        - pending_change_count
        - has_deployed
      properties:
        id:
          type: string
        name:
          type: string
        image:
          type: string
        container_name:
          type: string
        container_port:
          type: integer
          description: Container port the image listens on.
        host_port:
          type: integer
          description: >-
            Host port the service is published on. Omit to keep the service
            internal — reachable by other services on the uploy network and by
            nothing outside the machine, which is what a database wants. 80 and
            443 are rejected: they belong to the Uploy proxy. Ignored while the
            service has a domain, since that traffic goes through the proxy.
        server_id:
          type: string
        kind:
          type: string
          enum:
            - application
        project_id:
          type: string
        environment_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        pending_change_count:
          type: integer
          description: >
            How many changes are waiting to be deployed: the difference between
            the service's current configuration and the configuration its last
            successful deployment actually shipped, counted per field, per
            domain and per variable. Zero means the running container matches
            what the service says it is.

            A service with nothing to compare against — never deployed, or last
            deployed before Uploy recorded configurations — reports 1. It is
            pending, but there is no itemised answer to give.

            Derived per request, never stored.
        has_deployed:
          type: boolean
          description: >
            True once at least one deployment of this service has succeeded.
            Combined with has_pending_changes it distinguishes a service that
            will be created on its server from one that will be updated.
            Derived per request, never stored.
...